refact - refactoring the methods isEmailAndNameValidPersonalized, addToCart, defaultFunctionalityAddToCart and run, created simplifiedViewProducts method and was added exit option, all in ShoppingSystem class.
isEmailAndNameValidPersonalized: return a boolean value now if name length > 2 and email, with @gmail.com and email length > 10 that is length the string "@gmail.com".

addToCart: Modified only the case that now we are using the method simplifiedViewProducts to see what products are available.

defaultFunctionalityAddToCart: Was added only a verification if the option was back to menu.

run: Only welcome interface was added.

simplifiedViewProducts: Method created to see available products in a simplified way